src/authservice/pyproject.toml
[tool.poetry]
name = "authservice"
version = "0.1.0"
description = ""
authors = ["Anoop Rawat <anooprawat109@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.1"
autopep8 = "^1.6.0"
flake8 = "^4.0.1"
uvicorn = "^0.17.6"
python-keycloak = "^0.27.0"
pydantic = {extras = ["email"], version = "^1.9.0"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
